controls now kinda work if you set them. figured out that they were screwed up due to char overflow, but changing to int screws it all up
git-svn-id: svn+ssh://svn.code.sf.net/p/snowstorm/code/trunk/openglad@187 636a839e-d62e-0410-834d-a1d0aacdb804

+++ input.cpp
@@ -163,7 +163,7 @@ void key_int()
   //remember cli/sti
 }
 
-short query_key()
+int query_key()
 {
   return raw_key;
 }
